tests/test_cases/requirements.txt
# Fully resolved list of dependencies is required by Bazel.
pytest-metadata
pytest-env
psutil
testing-utils @ git+https://github.com/eclipse-score/testing_tools.git@v0.3.0
